#653 - convert footer accordion to a mixin
